storage: fold merge_cache into Cache
